requirements.txt
streamlit
google-genai
requests
beautifulsoup4
lxml
pandas
python-dateutil
pytrends
fredapi
feedparser
